Do name mangling on nucleus functions.
Since they are turned straight into symbols in the binary, it's really
easy to conflict with stuff from the standard library, and it would be
even worse once we start linking against other libraries.

Our name mangling scheme guarantees we won't clash with C functions, as
we use brackets, which are illegal in C identifiers.