Add a crippled form of closures.
Lambdas can now capture bindings from their environment, and read or
modify their values. They are crippled, however, as we make no attempt
at ensuring the values captured live as long as the closure does. For
example, if you return a lambda which refers to local values, when it
refers to them it will read garbage as the stack frame will be gone.

To fix this we need to do some form of escape analysis, and allocate
local variables on the heap instead of the stack where necessary.