Fix a bug with certain numeric/dynamic eq? comparisons.
When doing, for example (eq? x 0) (where x is of type ?), x would get
coerced to the same type as 0. 0 defaults to a i8, so this results in a
loss of information if the runtime value of x doesn't fit into an i8.
So, for example, (eq? x 0) would evaluate to true if x was dynamic and
had the value 256 or 512, as after the lossy conversion all the one bits
are gone.