Make a wrapper around rt_check_type taking fewer args.
The old version was designed to be called from C, and thus took extra
arguments for file, function, and line no for better diagnostics.
However, passing 0 for these inflated the call sequence, and
rt_check_type is called all over the place.

In terms of lines of assembly generated, this change reduces the size of
the compiler/stdlib by almost 10%.